Introduce DatabricksSqlWarehouseQueryExecutor
A new class to handle SQL statements
